+++ make/src/CocoaTech/CocoaTechFile/NTLaunchServices.m
@@ -110,7 +110,7 @@ extern OSStatus _LSCopyAllApplicationURLs(CFArrayRef * outURLs);
             result = [NSMutableArray arrayWithArray:(NSArray*)outURLs];
             
             // filter out .exe files
-            int i, cnt = [result count];
+            NSInteger i, cnt = [result count];
             NSURL *url;
             
             for (i=(cnt-1);i>=0;i--)
